+++ src/Git_Updater/Additions/Repo_List_Table.php
@@ -79,6 +79,7 @@ class Repo_List_Table extends WP_List_Table {
 			$option['primary_branch']  = ! empty( $option['primary_branch'] ) ? $option['primary_branch'] : 'master';
 			$option['release_asset']   = ! empty( $option['release_asset'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
 			$option['private_package'] = ! empty( $option['private_package'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
+			$option['uses_lite']       = ! empty( $option['uses_lite'] ) ? '<span class="dashicons dashicons-yes"></span>' : false;
 
 			$options[ $key ] = $option;
 		}
@@ -119,7 +120,8 @@ class Repo_List_Table extends WP_List_Table {
 		switch ( $column_name ) {
 			case 'release_asset':
 			case 'private_package':
-				return wp_kses_post( $item[ $column_name ] );
+			case 'uses_lite':
+				return empty( $item[ $column_name ] ) ? false : wp_kses_post( $item[ $column_name ] );
 			case 'uri':
 			case 'slug':
 			case 'primary_branch':
@@ -217,6 +219,7 @@ class Repo_List_Table extends WP_List_Table {
 			'primary_branch'  => esc_html__( 'Primary Branch', 'git-updater' ),
 			'release_asset'   => esc_html__( 'Release Asset', 'git-updater' ),
 			'private_package' => esc_html__( 'Private Package', 'git-updater' ),
+			'uses_lite'       => esc_html__( 'Lite Client', 'git-updater' ),
 			'type'            => esc_html__( 'Type', 'git-updater' ),
 		];